@@ -27,13 +27,14 @@ add_code_coverage_all_targets(EXCLUDE libraries aws-cpp-sdk googletest Integrati
 
 # This wasn't being picked up on mac, causes some symbol errors
 if(APPLE)
-	set(CMAKE_CXX_STANDARD 14)
+	set(CMAKE_CXX_STANDARD 20)
 	add_definitions(-DHAVE_STRLCAT)
 	# Do not use shared libraries for AWS SDK.
 	option(BUILD_SHARED_LIBS "Build shared libraries" OFF)
 	# Set BUILD_WITH_TESTS to OFF before building installer package for size optimization.
 	option(BUILD_WITH_TESTS "Enable testing" ON)
 else()
+	set(CMAKE_CXX_STANDARD 17)
 	# Use shared libraries for AWS SDK.
 	option(BUILD_SHARED_LIBS "Build shared libraries" ON)
 	# Set BUILD_WITH_TESTS to OFF before building installer package for size optimization.
@@ -112,6 +113,8 @@ add_compile_definitions		(	_SILENCE_TR1_NAMESPACE_DEPRECATION_WARNING
 								_MBCS
 								_CRT_SECURE_NO_DEPRECATE	
 								_USRDLL	
+								_SILENCE_CXX17_ITERATOR_BASE_CLASS_DEPRECATION_WARNING
+								_SILENCE_CXX17_CODECVT_HEADER_DEPRECATION_WARNING
 								# USE_SSL					
 							)
 
