@@ -100,7 +100,7 @@ const int delay_offset_3_1 = 0;
 const int delay_offset_3_2 = 180;
 const SQLSMALLINT single_col_name_length = 6;
 const SQLSMALLINT single_col_data_type = SQL_WVARCHAR;
-const SQLULEN single_col_column_size = 0;
+const SQLULEN single_col_column_size = 25;
 const SQLSMALLINT single_col_decimal_digit = 0;
 const SQLSMALLINT single_col_nullable = 2;
 const std::wstring single_row = L"1";
