@@ -104,8 +104,8 @@ void ESClearResult(ESResult* es_result) {
     delete es_result;
 }
 
-void ESClearQueue(void* es_conn) {
-     static_cast< ESCommunication* >(es_conn)->ClearQueue();
+void ESStopRetrieval(void* es_conn) {
+     static_cast< ESCommunication* >(es_conn)->StopResultRetrieval();
 }
 
 // This class provides a cross platform way of entering critical sections
