@@ -14,44 +14,51 @@
  *
  */
 
-#define QUEUE_CAPACITY 2
 #include "es_result_queue.h"
 
-ESResultQueue::ESResultQueue()
-    : m_queue_capacity(QUEUE_CAPACITY) {
+#include "es_types.h"
+
+ESResultQueue::ESResultQueue(unsigned int capacity)
+    : m_push_semaphore(capacity, capacity),
+      m_pop_semaphore(0, capacity) {
 }
 
 ESResultQueue::~ESResultQueue() {
-}
-
-std::reference_wrapper< ESResult > ESResultQueue::pop_front() {
-    std::reference_wrapper< ESResult > es_result = m_queue.front();
-    m_queue.pop();
-    return es_result;
-}
-
-void ESResultQueue::pop() {
-    m_queue.pop();
-}
-
-SQLRETURN ESResultQueue::push(std::reference_wrapper< ESResult > es_result) {
-    if (m_queue.size() < m_queue_capacity) {
-        m_queue.push(es_result);
-        return SQL_SUCCESS;
+    while (!m_queue.empty()) {
+        delete m_queue.front();
+        m_queue.pop();
     }
-    return SQL_ERROR;
-}
-
-bool ESResultQueue::empty() {
-    return m_queue.empty();
-}
-
-bool ESResultQueue::IsFull() {
-    return (m_queue.size() >= m_queue_capacity) ? true : false;
 }
 
 void ESResultQueue::clear() {
+    std::scoped_lock lock(m_queue_mutex);
     while (!m_queue.empty()) {
+        delete m_queue.front();
         m_queue.pop();
+        m_push_semaphore.release();
+        m_pop_semaphore.lock();
     }
+}
+
+bool ESResultQueue::pop(unsigned int timeout_ms, ESResult*& result) {
+    if (m_pop_semaphore.try_lock_for(timeout_ms)) {
+        std::scoped_lock lock(m_queue_mutex);
+        result = m_queue.front();
+        m_queue.pop();
+        m_push_semaphore.release();
+        return true;
+    }
+
+    return false;
+}
+
+bool ESResultQueue::push(unsigned int timeout_ms, ESResult* result) {
+    if (m_push_semaphore.try_lock_for(timeout_ms)) {
+        std::scoped_lock lock(m_queue_mutex);
+        m_queue.push(result);
+        m_pop_semaphore.release();
+        return true;
+    }
+
+    return false;
 }
