@@ -163,8 +163,7 @@ RETCODE SQL_API ESAPI_FreeStmt(HSTMT hstmt, SQLUSMALLINT fOption) {
     if (fOption == SQL_DROP) {
         ConnectionClass *conn = stmt->hdbc;
 
-        // Clear queue in case it holds next pages of results
-        ESClearQueue(conn->esconn);
+        ESStopRetrieval(conn->esconn);
 
         /* Remove the statement from the connection's statement list */
         if (conn) {
@@ -202,6 +201,8 @@ RETCODE SQL_API ESAPI_FreeStmt(HSTMT hstmt, SQLUSMALLINT fOption) {
     } else if (fOption == SQL_UNBIND)
         SC_unbind_cols(stmt);
     else if (fOption == SQL_CLOSE) {
+        ESStopRetrieval(stmt->hdbc->esconn);
+
         /*
          * this should discard all the results, but leave the statement
          * itself in place (it can be executed again)
